Brings in the test overhaul (#139), the Ruff/pip-audit setup, the CI
workflows, and the decimal-input fix (#133).

Conflicts, and how they were taken:

Deleted files. main lint-fixed six modules this branch had already
deleted — AccountSettingsPage, AddFriendModal, GroupDetailPage,
ManageGuestModal, ManageMemberModal, SendFriendRequestModal. Every change
was an unused catch binding or an `any`; all six are superseded by the
redesign and nothing in main's new tests references them, so they stay
deleted.

Import blocks. main sorted them with Ruff's isort; this branch had added
names to the unsorted originals. Took main's ordering with our additions
folded in: tabs in main.py, UniqueConstraint and Numeric in models.py,
re and List in schemas.py, the display helpers in balances.py.

App.tsx. main lint-fixed the old Dashboard-based app that this branch
replaced wholesale; ours kept.

SettleUpModal. main introduced a typed ExpensePayload and renamed the
alert state; this branch had added an amount guard and passes its payload
inline. Kept the guard and the inline call, adopted the type — which
promptly caught that both settle paths could send payer_id: undefined.
Now guarded.

Two things carried in from main were already broken there, and are fixed
here so this branch is green at every commit:

  * useSplitDetails.test.ts asserted the eager `parseFloat(value) || 0`
    that #133 deliberately removed. #133 is right — parsing on every
    keystroke makes a decimal impossible to type, since "0." collapses
    to 0 before the digits arrive. The tests now assert the raw-text
    contract.

  * That same change let unparseable text reach the split calculators.
    The field is type="text", so "abc" is typeable; parseFloat turned it
    into NaN, which rode into amount_owed and serialises as null in the
    request body — and under SHARES one bad entry zeroed every
    participant, not just its own. Parsing now goes through
    parseSplitValue, which floors anything unparseable at 0. The
    guarantee the deleted assertions protected is kept, just at the
    layer that now owns it.

Backend: 549 pass, ruff clean. Frontend: 348 pass, tsc clean, lint:ci
clean, build succeeds.

+++ backend/schemas.py
@@ -1,8 +1,8 @@
 import re
-
-from pydantic import BaseModel, EmailStr, field_validator, Field
-from typing import Optional, Dict, List, Literal
 from datetime import datetime
+from typing import Dict, List, Literal, Optional
+
+from pydantic import BaseModel, EmailStr, Field, field_validator
 
 from utils.currency import VALID_CURRENCIES
 
@@ -31,7 +31,7 @@ class ExpenseSplitBase(BaseModel):
     is_guest: bool = False
     amount_owed: int
     percentage: Optional[int] = None
-    shares: Optional[int] = None
+    shares: Optional[float] = None
 
 # Expense guest schemas (for non-group expenses)
 class ExpenseGuestCreate(BaseModel):
@@ -68,7 +68,7 @@ class ItemAssignment(BaseModel):
 class ItemSplitDetail(BaseModel):
     amount: Optional[int] = None  # For EXACT split (in cents)
     percentage: Optional[float] = None  # For PERCENTAGE split (0-100)
-    shares: Optional[int] = None  # For SHARES split
+    shares: Optional[float] = None  # For SHARES split
 
 class ExpenseItemCreate(BaseModel):
     description: str = Field(..., max_length=200)
@@ -143,7 +143,7 @@ class ExpenseSplitDetail(BaseModel):
     is_guest: bool = False
     amount_owed: int
     percentage: Optional[int] = None
-    shares: Optional[int] = None
+    shares: Optional[float] = None
     user_name: str
 
     class Config:
@@ -476,7 +476,7 @@ class FriendExpenseWithSplits(ExpenseWithSplits):
 
 
 # Profile Management and Password Recovery Schemas
-from datetime import datetime
+
 
 class PasswordChangeRequest(BaseModel):
     """Request to change password (requires current password)"""
